NodeUtility.FastForwardAsync
